+++ test/test_helper.rb
@@ -23,8 +23,10 @@ $VERBOSE = nil
 
 # Load the Redmine helper
 require File.expand_path('../../../test/test_helper', __dir__)
+require File.expand_path('../../../test/application_system_test_case', __dir__)
 require_relative 'load_fixtures'
 require_relative 'authenticate_user'
+require_relative 'enumerations'
 require_relative 'project_type_creator'
 require_relative 'inheritated_spreadsheets'
 require_relative 'test_object_creators'