+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 * wikitoolbar to spreadsheet result row comment field
 * prefilled spreadsheet result row data fields
+* colored badges for custom field enumerations
 
 ## 0.1.1 - 2021-07-13